+++ src/main/java/ru/practicum/shareit/booking/BookingStorage.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.model.EnumStatusBooking;
+
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -38,7 +39,6 @@ public interface BookingStorage extends JpaRepository<Booking, Long> {
     Booking findFirstByItem_IdAndStartAfterAndStatusOrderByStartAsc(Long id, LocalDateTime start, EnumStatusBooking status);
 
     Booking findFirstByItem_IdAndStartBeforeAndStatusOrderByStartDesc(Long id, LocalDateTime start, EnumStatusBooking status);
-
 
 
 }
