+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -1,67 +1,19 @@
 package ru.practicum.shareit.user;
 
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exceptions.NotFoundParametrException;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
-import javax.transaction.Transactional;
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@Slf4j
-public class UserService {
+public interface UserService {
 
-    private final UserStorage userStorage;
+    UserDto createUser(User user);
 
-    @Autowired
-    public UserService(UserStorage userStorage) {
-        this.userStorage = userStorage;
-    }
+    UserDto patchUser(User user, Long userId);
 
-    public UserDto createUser(User user) {
-        return UserMapper.toUserDto(userStorage.save(user));
-    }
+    UserDto getUserById(Long userId);
 
-    @Transactional
-    public UserDto patchUser(User user, Long userId) {
-        User patchingUser = userStorage.findById(userId).get();
-        if (user.getName() != null) {
-            patchingUser.setName(user.getName());
-            log.info("у пользователя с id {} заменено имя на {}", userId, patchingUser.getName());
-        }
-        if (user.getEmail() != null) {
-            patchingUser.setEmail(user.getEmail());
-            log.info("у пользователя с id {} заменено имя на {}", userId, patchingUser.getEmail());
-        }
+    List<UserDto> getAllUsers();
 
-        return UserMapper.toUserDto(patchingUser);
-    }
-
-    public UserDto getUserById(Long userId) {
-        var rez = userStorage.findById(userId);
-        if (rez.isEmpty()) {
-            throw new NotFoundParametrException(String.format("Отсутствует запрашиваемый пользователь. userId= %d",
-                    userId));
-        }
-        return UserMapper.toUserDto(rez.get());
-    }
-
-    public List<UserDto> getAllUsers() {
-        return userStorage.findAll().stream()
-                .map(UserMapper::toUserDto)
-                .collect(Collectors.toList());
-    }
-
-    public void deleteUserById(Long userId) {
-        userStorage.deleteById(userId);
-    }
-
-    private List<User> isUsersEmailDuplicate(User user) {
-        return userStorage.findByEmailContainingIgnoreCase(user.getEmail());
-    }
-
+    void deleteUserById(Long userId);
 }
